Kevin[15min]: start restructuring backend
