+++ back_end/server.go
@@ -3,8 +3,6 @@ package main
 import (
 	"database/sql"
 	"fmt"
-	"net/http"
-	"strconv"
 
 	"github.com/gin-contrib/cors"
 	"github.com/gin-gonic/gin"
@@ -19,11 +17,6 @@ const PORT_NO = ":8080"
 // Then call these queries in the route handler funcs to test functionality of queries
 // Move route handlers and maybe queries to their own files (ex. routes.go, queries.go)
 // TODO Get postgres variable into all handlers
-
-type User struct {
-	ID   string `json:"id"`
-	Name string `json:"name"`
-}
 
 // Connecting PostgreSQL
 const (
@@ -44,6 +37,7 @@ func main() {
 	server := gin.Default()
 	server.ForwardedByClientIP = true
 	server.SetTrustedProxies([]string{"127.0.0.1", "client"}) // Add any other needed IPs
+	setupRoutes(server)
 	config := cors.DefaultConfig()
 
 	config.AllowOrigins = []string{"http://localhost:3000", "http://localhost:8080", "*"}
@@ -62,8 +56,6 @@ func main() {
 		if err != nil {
 			panic(err)
 		}
-		fmt.Printf("Go Postgres! %v\n", postgres)
-		// initialize(postgres)
 		defer postgres.Close()
 		ctx.Set("postgres", postgres)
 
@@ -89,12 +81,7 @@ func main() {
 	// Route handlers for API endpoints:
 	server.Handler()
 	// server.GET("/postgresTest", testPostgres)
-	server.GET("/ping", pong)
-	server.PUT("/num")
 
-	server.GET("/num/:num1/:num2", sum)
-	server.GET("/testInsert/:val", testInsertHandler)
-	server.POST("/user", double)
 	server.Run(PORT_NO)
 }
 
@@ -219,39 +206,3 @@ func testInsertHandler(ctx *gin.Context) {
 // 		})
 // 	}
 // }
-
-func double(ctx *gin.Context) {
-	var inputUser User
-	if err := ctx.ShouldBindJSON(&inputUser); err != nil {
-		ctx.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
-		return
-	}
-
-	newID := inputUser.ID + inputUser.ID
-	newName := inputUser.Name + inputUser.Name
-
-	responseUser := User{
-		ID:   newID,
-		Name: newName,
-	}
-
-	ctx.JSON(http.StatusOK, responseUser)
-}
-
-func pong(ctx *gin.Context) {
-	ctx.JSON(http.StatusOK, gin.H{
-		"message": "pong",
-	})
-}
-
-func sum(ctx *gin.Context) {
-	num1, err1 := strconv.Atoi(ctx.Param("num1"))
-	num2, err2 := strconv.Atoi(ctx.Param("num2"))
-	if err1 != nil {
-		ctx.JSON(http.StatusBadRequest, gin.H{"error": err1.Error()})
-	} else if err2 != nil {
-		ctx.JSON(http.StatusBadRequest, gin.H{"error": err2.Error()})
-	}
-	sum := num1 + num2
-	ctx.String(http.StatusOK, strconv.Itoa(sum))
-}
